refactor: 패키지 구조 변경 #17
- components 파일들 패키지로 묶어 관리
- 안쓰는 컴포넌트 삭제
